Replaced "libusb" backend specifier with "*" wildcard
This allows other backends, such as the Cypress backend to be used
by libbladeRF.

Note that in the case where multiple devices are present on different
backends, it is possible for each device to be "instance 0" on each
backend. In the future it might be better to decouple the device ID
number the user provides from the actual instance number passed to the
device.